examples: set crtc/connector properties
The behind-the-scenes manipulation of the per-plane DRM properties
needs to be augmented with some singleton property assignments
too (what CRTC should the connector be driving? what mode should
be used? is the CRC on?).

Also use DRM_MODE_ATOMIC_ALLOW_MODESET in the live commit in
case the new configuration doesn't match with whatever set of
modes and color formats the previous DRM master happened to use.