Disable plane reset logging
Logigng the list of planes being disabled at the beginning of an apply does not
really give useful information.

Planes are disabled (and by that become in theory ready for layer-assignment)
if they do not have a layer associated at the moment or if they are compatible
with the output.

That is uninteresting since in this list the subset of planes that are
compatible with the output stay constant. And while the subset of planes being
reset not compatible with the output might vary from apply to apply it is also
of less interest since by definition they are not compatible with what the
apply is about.